Update README.md
I am a student applying the 19 fall master program in the US, and at the same time, I need to complete a graduation project with my partner.
Therefore, I choose to join the datawhale study group to better and quicker finish the project by learning python.
Moreover, it is not a good march for me, since I already got four rejection this month including my dream school. And What is worse? I did nothing on my study in this month. So, I must do something to make me feel better in the next month.
People need to move on, and success will only is to someone who jumps out the failure and is prepared for the next time.
Hope I can completely finish this project and really get something.